rootSaga, store and apiCode moved outside the App folder into core folder
