+++ src/core/rootSaga.js
@@ -1,5 +1,5 @@
 import { all } from "redux-saga/effects";
-import { popularMoviesSaga } from "../../features/MovieList/popularMoviesSaga";
+import { popularMoviesSaga } from "../features/MoviesList/popularMoviesSaga";
 
 export default function* rootSaga() {
   yield all([popularMoviesSaga()]);
