+++ src/core/store.js
@@ -1,7 +1,7 @@
 import createSagaMiddleware from "@redux-saga/core";
 import { configureStore } from "@reduxjs/toolkit";
-import popularMoviesReducer from "../../features/MovieList/popularMoviesSlice";
-import rootSaga from "../App/rootSaga";
+import popularMoviesReducer from "../features/MovieList/popularMoviesSlice";
+import rootSaga from "./rootSaga";
 
 const sagaMiddleware = createSagaMiddleware();
